+++ CHANGELOG.md
@@ -4,3 +4,7 @@
 
 * Made Shtern
 
+## 2.0.0
+
+* Made default mask character an empty string because it is more secure
+* Added optional `prompt` parameter